[FIX] account_payment: Wrong UI tab mention in error
Steps:
- Install "invoicing"
- Install payment provider (Stripe)
- Enable the provider (test mode)
- Go to apps app
- Find the provider module
- Uninstall it

Actual result:
- I can't uninstall stripe due to "outgoing" data
- Error message doesn't mention the right tab

Expected result:
- I can't uninstall stripe due to "incoming" data
- Error message mention the right tab

X-original-commit: 5b56d4e8c8f99b982c4cb018874f1e30824cd19d

+++ addons/account_payment/models/payment_provider.py
@@ -109,5 +109,7 @@ class PaymentProvider(models.Model):
         # If the payment method is used by any payments, we block the uninstallation of the module.
         if self._check_existing_payment(payment_method):
             raise UserError(_("You cannot uninstall this module as payments using this payment method already exist."))
+        if self._check_existing_payment_method_lines(payment_method):
+            raise UserError(_("To uninstall this module, please remove first the corresponding payment method line in the outgoing payments tab defined on the bank journal."))
         super()._remove_provider(code)
         payment_method.unlink()
